Removed all non-schema tests
